third_party/blink/web_tests/platform/linux-chrome/external/wpt/clipboard-apis/async-custom-formats-write-read.tentative.https-expected.txt

This is a testharness.js-based test.
[FAIL] Verify write and read clipboard given 2 platform-neutral custom format inputs
  assert_equals: expected "input data 1" but got ""
Harness: the test ran to completion.
